backend/requirements.txt
Flask==2.3.0
twilio==7.21.0
python-dotenv==1.0.0
